// admin-ui/src/app/api/admin-api.service.generated.ts
export class PostDTO implements IPostDTO {
    id?: string;
    name!: string | undefined;
    slug!: string | undefined;
    description?: string | undefined;
    thumbnail?: string | undefined;
    viewCount?: number;
    dateCreated?: Date;
    categorySlug!: string | undefined;
    categoryName!: string | undefined;
    authorUserName?: string | undefined;
    authorName?: string | undefined;
    status?: PostStatus;
    isPaid?: boolean;
    royaltyAmount?: number;
    paidDate?: Date | undefined;
    categoryId?: string;
    content?: string | undefined;
    authorUserId?: string;
    source?: string | undefined;
    tags?: string | undefined;
    seoDescription?: string | undefined;
    dateModified?: Date | undefined;

    constructor(data?: IPostDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.slug = _data["slug"];
            this.description = _data["description"];
            this.thumbnail = _data["thumbnail"];
            this.viewCount = _data["viewCount"];
            this.dateCreated = _data["dateCreated"] ? new Date(_data["dateCreated"].toString()) : undefined as any;
            this.categorySlug = _data["categorySlug"];
            this.categoryName = _data["categoryName"];
            this.authorUserName = _data["authorUserName"];
            this.authorName = _data["authorName"];
            this.status = _data["status"];
            this.isPaid = _data["isPaid"];
            this.royaltyAmount = _data["royaltyAmount"];
            this.paidDate = _data["paidDate"] ? new Date(_data["paidDate"].toString()) : undefined as any;
            this.categoryId = _data["categoryId"];
            this.content = _data["content"];
            this.authorUserId = _data["authorUserId"];
            this.source = _data["source"];
            this.tags = _data["tags"];
            this.seoDescription = _data["seoDescription"];
            this.dateModified = _data["dateModified"] ? new Date(_data["dateModified"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): PostDTO {
        data = typeof data === 'object' ? data : {};
        let result = new PostDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["slug"] = this.slug;
        data["description"] = this.description;
        data["thumbnail"] = this.thumbnail;
        data["viewCount"] = this.viewCount;
        data["dateCreated"] = this.dateCreated ? this.dateCreated.toISOString() : undefined as any;
        data["categorySlug"] = this.categorySlug;
        data["categoryName"] = this.categoryName;
        data["authorUserName"] = this.authorUserName;
        data["authorName"] = this.authorName;
        data["status"] = this.status;
        data["isPaid"] = this.isPaid;
        data["royaltyAmount"] = this.royaltyAmount;
        data["paidDate"] = this.paidDate ? this.paidDate.toISOString() : undefined as any;
        data["categoryId"] = this.categoryId;
        data["content"] = this.content;
        data["authorUserId"] = this.authorUserId;
        data["source"] = this.source;
        data["tags"] = this.tags;
        data["seoDescription"] = this.seoDescription;
        data["dateModified"] = this.dateModified ? this.dateModified.toISOString() : undefined as any;
        return data;
    }
}

export interface IPostDTO {
    id?: string;
    name: string | undefined;
    slug: string | undefined;
    description?: string | undefined;
    thumbnail?: string | undefined;
    viewCount?: number;
    dateCreated?: Date;
    categorySlug: string | undefined;
    categoryName: string | undefined;
    authorUserName?: string | undefined;
    authorName?: string | undefined;
    status?: PostStatus;
    isPaid?: boolean;
    royaltyAmount?: number;
    paidDate?: Date | undefined;
    categoryId?: string;
    content?: string | undefined;
    authorUserId?: string;
    source?: string | undefined;
    tags?: string | undefined;
    seoDescription?: string | undefined;
    dateModified?: Date | undefined;
}
// ...
export enum PostStatus {
    _0 = 0,
    _1 = 1,
    _2 = 2,
    _3 = 3,
}